+++ src/cli/Microsoft.DotNet.UpgradeAssistant.Cli/ConsoleRunner.cs
@@ -5,6 +5,10 @@ using System;
 using System.Collections.Generic;
 using System.Threading;
 using System.Threading.Tasks;
+using Autofac;
+using Autofac.Extensions.DependencyInjection;
+using Microsoft.DotNet.UpgradeAssistant.Extensions;
+using Microsoft.DotNet.UpgradeAssistant.Telemetry;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Hosting;
 using Microsoft.Extensions.Logging;
@@ -20,16 +24,19 @@ namespace Microsoft.DotNet.UpgradeAssistant.Cli
         private readonly ILogger _logger;
         private readonly IHostApplicationLifetime _lifetime;
         private readonly ErrorCodeAccessor _errorCode;
+        private readonly ITelemetry _telemetry;
 
         public ConsoleRunner(
             IServiceProvider services,
             IHostApplicationLifetime lifetime,
             ErrorCodeAccessor errorCode,
+            ITelemetry telemetry,
             ILogger<ConsoleRunner> logger)
         {
             _lifetime = lifetime ?? throw new ArgumentNullException(nameof(lifetime));
             _services = services ?? throw new ArgumentNullException(nameof(services));
             _errorCode = errorCode ?? throw new ArgumentNullException(nameof(errorCode));
+            _telemetry = telemetry ?? throw new ArgumentNullException(nameof(telemetry));
             _logger = logger ?? throw new ArgumentNullException(nameof(logger));
         }
 
@@ -40,8 +47,19 @@ namespace Microsoft.DotNet.UpgradeAssistant.Cli
             {
                 _logger.LogDebug("Configuration loaded from context base directory: {BaseDirectory}", AppContext.BaseDirectory);
 
-                await RunStartupAsync(token);
-                await RunCommandAsync(token);
+                using var scope = _services.GetAutofacRoot().BeginLifetimeScope(builder =>
+                {
+                    foreach (var extension in _services.GetRequiredService<IEnumerable<ExtensionInstance>>())
+                    {
+                        var services = new ServiceCollection();
+                        extension.AddServices(services);
+                        builder.Populate(services);
+                    }
+                });
+
+                await RunStartupAsync(scope.Resolve<IEnumerable<IUpgradeStartup>>(), token);
+
+                await scope.Resolve<IAppCommand>().RunAsync(token);
 
                 _errorCode.ErrorCode = ErrorCodes.Success;
             }
@@ -49,27 +67,29 @@ namespace Microsoft.DotNet.UpgradeAssistant.Cli
             {
                 _logger.LogError("{Message}", e.Message);
                 _errorCode.ErrorCode = ErrorCodes.UpgradeError;
+                _telemetry.TrackException(e);
             }
             catch (OperationCanceledException)
             {
                 _logger.LogTrace("A cancellation occurred");
+                _errorCode.ErrorCode = ErrorCodes.Canceled;
             }
             catch (Exception e)
             {
-                _logger.LogError(e, "Unexpected error during upgrade.");
+                _logger.LogError(e, "Unexpected error");
                 _errorCode.ErrorCode = ErrorCodes.UnexpectedError;
+                _telemetry.TrackException(e);
             }
             finally
             {
+                _telemetry.TrackEvent("exited", measurements: new Dictionary<string, double> { { "Exit Code", _errorCode.ErrorCode } });
+
                 _lifetime.StopApplication();
             }
         }
 
-        private async Task RunStartupAsync(CancellationToken token)
+        private static async Task RunStartupAsync(IEnumerable<IUpgradeStartup> startups, CancellationToken token)
         {
-            using var scope = _services.CreateScope();
-            var startups = scope.ServiceProvider.GetRequiredService<IEnumerable<IUpgradeStartup>>();
-
             foreach (var startup in startups)
             {
                 if (!await startup.StartupAsync(token))
@@ -77,13 +97,6 @@ namespace Microsoft.DotNet.UpgradeAssistant.Cli
                     throw new UpgradeException($"Failure running start up action {startup.GetType().FullName}");
                 }
             }
-        }
-
-        private async Task RunCommandAsync(CancellationToken token)
-        {
-            using var scope = _services.CreateScope();
-            var command = scope.ServiceProvider.GetRequiredService<IAppCommand>();
-            await command.RunAsync(token);
         }
 
         public Task StopAsync(CancellationToken cancellationToken) => Task.CompletedTask;